Refactor onnxmltools common folder into onnxtk (#268)
* move onnxmltools/convert/common -> onnxtk/common

* fixing folder structure dependencies in common + convert

* removing folder structure and using module imports instead

* Replacing all converter dependencies on ...common and ....common with onnxutils.onnxtk.common

* update folder structure dependency to extract model producer, domain

* fixing last import disparities in onnxutils and test_utils

* replacing "onnxutils.onnxtk" with "onnxtk"

* Adding onnxtk to the module requirements

* Using onnxmltools.proto, adding updated description for onnxtk

* testing the PR version of onnxtk in conjunction with PR for onnxmltools

* testing once more on win32 yml CI script

* CI build testing

* fixing imports for onnxtk files -- using onnx model proto

* CI Build -- fixing no onnxruntime error

* updating dependency build instructions for onnxtk for linux yml file

* fixing one last onnxmltools onnx_proto dependency from onnxtk in optimizer

* replacing data_types imports, fixing test imports, adding "import onnxtk" in coreml ops

* fixing remaining imports of ..common with onnxtk.common

* XGBoost's ..common utilities are not be placed in onnxtk

* 1 more missed XGBoost ..common utility

* fixing requirements + domain for "onnxtk" as "onnxml" to unbreak CI build

* removing "import onnxtk" from coreml ops

* resolve onnxtk domain and test case data domain for baseline ImageNet conversion model

* remove onnxruntime 0.1.4 installation

* removing metadata_props duplicate in onnxmltools_utils

* removing duplicate optimizer

* removed opset version argument as it wasn't being used in optimize_onnx and was breaking backwards compatibility

* updating windows CI build with onnxruntime installation

* remove onnxruntime installation in this PR, revert to previous version

* wrapper for onnx common utilities

* using convert/common imports with onnxtk module wrapper for better backwards compatibility

* Updating wrapping functionality for onnxmltools backwards compatibility

* fixing import folder structure error for sparkml utils

* importing _check_has_attr explicitly

* updating one last common.utils libsvm dependency

* resolving review comment pertaining to local clone in CI build

* ensuring backwards compatibility for ONNX graph optimizer

* fixing folder structure for onnxtk install

* finding root folder to install onnxtk from

* restoring dependencies to onnxmltools using wrapper API, deprecating float16_converter

* Revising tutorial documentation for onnxtk change

* reverting test dependency to onnxmltools, no onnxtk depency on xmltools

* reverting last two test dependencies to use wrappers

* fixing typo in onnx graph optimizer
