+++ docs/index.rst
@@ -53,7 +53,7 @@ to automatically check every converter with
     # Convert into ONNX format with onnxmltools
     from onnxmltools import convert_sklearn
     from onnxmltools.utils import save_model
-    from onnxmltools.convert.common.data_types import FloatTensorType
+    from onnxtk.common.data_types import FloatTensorType
     initial_type = [('float_input', FloatTensorType([1, 4]))]
     onx = convert_sklearn(clr, initial_types=initial_type)
     save_model(onx, "rf_iris.onnx")
