+++ onnxmltools/convert/common/__init__.py
@@ -3,3 +3,5 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
+
+# ONNXMLTools common code has been refactored into onnxtk.
