+++ onnxmltools/convert/sparkml/shape_calculators/LogisticRegression.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from onnxmltools.convert.common.data_types import Int64TensorType, DictionaryType, SequenceType, FloatTensorType
+from ...common.data_types import Int64TensorType, DictionaryType, SequenceType, FloatTensorType
 from ...common._registration import register_shape_calculator
 from ...common.utils import check_input_and_output_numbers, check_input_and_output_types
