+++ onnxmltools/convert/sparkml/shape_calculators/VectorAssembler.py
@@ -6,7 +6,7 @@
 
 from ...common._registration import register_shape_calculator
 from ...common.utils import check_input_and_output_numbers
-from onnxmltools.convert.common.data_types import *
+from ...common.data_types import *
 
 def calculate_vector_assembler_shapes(operator):
     check_input_and_output_numbers(operator, output_count_range=1)
