+++ onnxmltools/convert/sparkml/utils.py
@@ -1,7 +1,7 @@
 '''
 Utility functions for Spark ML to Onnx conversion intended for the end user mainly
 '''
-from onnxmltools.convert.common.data_types import StringTensorType, Int64TensorType, FloatTensorType
+from ..common.data_types import StringTensorType, Int64TensorType, FloatTensorType
 
 
 def buildInitialTypesSimple(dataframe):
@@ -36,4 +36,3 @@ def buildInputDictSimple(dataframe):
         else:
             result[field.name] = dataframe.select(field.name).toPandas().values.astype(numpy.float32)
     return result
-
