+++ onnxmltools/utils/__init__.py
@@ -10,7 +10,7 @@ from .main import save_text
 from .main import set_model_version
 from .main import set_model_domain
 from .main import set_model_doc_string
-from .metadata_props import add_metadata_props, set_denotation
+from onnxtk.metadata_props import add_metadata_props, set_denotation
 from .visualize import visualize_model
 from .float16_converter import convert_float_to_float16
 from .tests_helper import dump_data_and_model
