+++ requirements.txt
@@ -4,3 +4,4 @@ protobuf
 onnx
 skl2onnx
 keras2onnx
+onnxtk
