+++ tests/xgboost/test_xgboost_unpickle_06.py
@@ -24,4 +24,3 @@ class TestXGBoostUnpickle06(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
-
